refactor: 파일 디렉토리 변경 및 변수 명 수정
